+++ Changelog.md
@@ -1,10 +1,11 @@
 ## v0.9.2 - 2015 February
-- Support for `put`, `head`, `patch`, `trace` and `delete` HTTP methods
+- Support for `options`, `put`, `head`, `patch`, `trace` and `delete` HTTP methods
 - Added `app.put`
 - Added `app.head`
 - Added `app.patch`
 - Added `app.delete`
 - Added `app.trace`
+- Added `app.options`
 - Fixed body parsing mechanism. Every route's body is parsed that has a Content-Type or Transfer-Encoding Header. According to rfc2616-sec4